pyproject.toml
[tool.poetry]
name = "acronomy"
version = "0.1.0"
description = ""
authors = ["Lukas Winkler <git@lw1.at>"]

[tool.poetry.dependencies]
python = "^3.8"
django = {extras = ["argon2"], version = "^3.0.6"}
psycopg2 = "^2.8.5"
django-simple-history = "^2.10.0"
markdown = "^3.2.2"
django-debug-toolbar = "^2.2"
djangorestframework = "^3.11.0"
requests = "^2.23.0"
ads = "^0.12.3"
Pillow = "^7.1.2"

[tool.poetry.dev-dependencies]
bpython = "^0.19"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
